- renamed namespace for Extension classes as suggested by Rider - renamed exceptionPathRegex to KoanExceptionPathRegex better clarify context - Updated Regex object to pattern match based on the Koan naming scheme starting with “About” -Output should now only show the relevant Koans Excercise
