+++ Engine/Extensions/ExceptionExtensions.cs
@@ -1,15 +1,16 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 
-namespace System
+namespace DotNetKoans.Engine.Extensions
 {
     public static class ExceptionExtensions
     {
-        private static readonly Regex exceptionPathRegex = new Regex(@"(\w*\.cs:line \d*)");
+        private static readonly Regex KoanExceptionPathRegex = new(@"(About\w*\.cs:line \d*)");
 
         public static IEnumerable<string> GetStackTracePaths(this Exception exception) =>
 
-            exceptionPathRegex.Matches(exception.StackTrace).Select(x => x.Value);
+            KoanExceptionPathRegex.Matches(exception.StackTrace).Select(x => x.Value);
     }
 }
